Move ListViewSimplified
