+++ packages/client/src/views/Settings/items/Name.tsx
@@ -10,7 +10,7 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import * as React from 'react'
-import { ListViewItemSimplified } from '@opencrvs/components/lib/interface'
+import { ListViewItemSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { useIntl } from 'react-intl'
 import { userMessages, buttonMessages } from '@client/i18n/messages'
 import {
