+++ packages/client/src/views/Settings/items/PIN.tsx
@@ -10,7 +10,7 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import * as React from 'react'
-import { ListViewItemSimplified } from '@opencrvs/components/lib/interface'
+import { ListViewItemSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { constantsMessages, buttonMessages } from '@client/i18n/messages'
 import { useIntl } from 'react-intl'
 import {
