+++ packages/client/src/views/SysAdmin/Config/Application/Tabs/GeneralProperties/NIDNumPattern.tsx
@@ -20,7 +20,7 @@ import {
   Value
 } from '@client/views/SysAdmin/Config/Application/Components'
 import { IStoreState } from '@client/store'
-import { ListViewItemSimplified } from '@opencrvs/components/lib/interface'
+import { ListViewItemSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { Toast, NOTIFICATION_TYPE } from '@opencrvs/components/lib/Toast'
 import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
 import { GeneralActionId } from '@client/views/SysAdmin/Config/Application'
