+++ packages/client/src/views/SysAdmin/Performance/utils.tsx
@@ -22,10 +22,8 @@ import endOfMonth from 'date-fns/endOfMonth'
 import React from 'react'
 import { getPercentage } from '@client/utils/data-formatting'
 import { FormattedNumber, IntlShape } from 'react-intl'
-import {
-  ListViewSimplified,
-  ISearchLocation
-} from '@opencrvs/components/lib/interface'
+import { ISearchLocation } from '@opencrvs/components/lib/interface'
+import { ListViewSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { constantsMessages } from '@client/i18n/messages'
 import { messages as statusMessages } from '@client/i18n/messages/views/registrarHome'
 import { colors } from '@opencrvs/components/lib/colors'
