+++ packages/components/src/ListViewSimplified/ListViewSimplified.stories.tsx
@@ -15,10 +15,10 @@ import {
   ListViewItemSimplified
 } from './ListViewSimplified'
 import React from 'react'
-import { ToggleMenu } from '../../ToggleMenu'
-import { VerticalThreeDots, AvatarSmall, AvatarLarge } from '../../icons'
-import { LinkButton, PrimaryButton } from '../../buttons'
-import { Pill } from '../../Pill'
+import { ToggleMenu } from '../ToggleMenu'
+import { VerticalThreeDots, AvatarSmall, AvatarLarge } from '../icons'
+import { LinkButton, PrimaryButton } from '../buttons'
+import { Pill } from '../Pill'
 import styled from 'styled-components'
 
 const Template: Story = ({ children, ...args }) => {
