+++ packages/components/src/interface/index.ts
@@ -13,7 +13,6 @@ export * from './Header'
 export * from './Tabs'
 export * from './ActionPageLight'
 export * from './SelectGroup'
-export * from './ListViewSimplified/ListViewSimplified'
 export * from './DataTable/SortAndFilter'
 export * from './SectionDrawer'
 export * from './SearchTool'
